v0.1.9

+++ i.sh
@@ -8,28 +8,44 @@ RELEASES_URL="https://github.com/docsion/rfsh/releases"
 FILE_BASENAME="runflow_sh"
 LATEST="$(curl -sf https://raw.githubusercontent.com/docsion/rfsh/main/latest)"
 
-test -z "$VERSION" && VERSION="$LATEST"
+if [[ -z "$VERSION" ]]; then
+  VERSION="$LATEST"
+fi
 
-test -z "$VERSION" && {
-	echo "[RFSH] Unable to get runflow version." >&2
- 	echo
-	exit 1
-}
+if [[ -z "$VERSION" ]]; then
+  echo "[RFSH] Unable to get runflow version." >&2
+  echo
+  exit 1
+fi
 
+# Download
 TMP_DIR="$(mktemp -d)"
 trap "rm -rf \"$TMP_DIR\"" EXIT INT TERM
 OS="$(uname -s)"
 ARCH="$(uname -m)"
 TAR_FILE="${FILE_BASENAME}_${OS}_${ARCH}.tar.gz"
+echo "[RFSH] Downloading runflow $VERSION..."
+echo
+n=0
+until [ "$n" -ge 3 ]; do
+  curl -C - --retry 3 -sfLo "$TMP_DIR/$TAR_FILE" "$RELEASES_URL/download/$VERSION/$TAR_FILE" && break
+  n=$((n+1))
+  sleep 15
+done
+if [[ ! -a "$TMP_DIR/$TAR_FILE" ]]; then
+  echo '[RFSH] Failed to download runflow'
+exit
+fi
 
-(
-	cd "$TMP_DIR"
-	echo "[RFSH] Downloading runflow $VERSION..."
- 	echo
-	curl -sfLO "$RELEASES_URL/download/$VERSION/$TAR_FILE"
-)
-
+# Install
 tar -xf "$TMP_DIR/$TAR_FILE" -C "$TMP_DIR"
 mv "$TMP_DIR/runflow_sh" ./runflow
-echo "[RFSH] Done" $(pwd)/runflow
+echo "[RFSH] Runflow has been installed"
+echo
+
+# Launch
+echo "[RFSH] Launch rfsh" 
+echo
+echo $(pwd)/runflow "$@"
+$(pwd)/runflow "$@"
 echo
